Moved SyntaxHighlighter initialization
#227
* Moved initialization of SyntaxHighlighter into its constructor.
